refactor(circuit): Fix circuit entity public api
- Fixed circuit entity public api's
