+++ packages/client/__tests__/e2e/app/circuits/page.test.ts
@@ -2,7 +2,7 @@ import { expect } from '@playwright/test';
 import { test } from '~tests-utils/e2e/server/MockApiTest';
 import { setupServer, closeServer } from '~tests-utils/e2e/server/MockFastifyServer';
 
-import { URLS } from '~entities/circuit/api/urls';
+import { CIRCUIT_URLS } from '~entities/circuit/api';
 
 import { CircuitsMock } from '~mocks/entities/circuit/Circuit.mock';
 
@@ -14,7 +14,7 @@ test.afterEach(async ({ server }) => {
 
 test('render circuits list', async ({ page, server }) => {
     await setupServer(server, {
-        url: URLS.index,
+        url: CIRCUIT_URLS.index,
         method: 'GET',
         handler: function (_, reply) {
             reply.send({
@@ -35,7 +35,7 @@ test('should open circuit page after link click', async ({ page, server }) => {
     await setupServer(
         server,
         {
-            url: URLS.index,
+            url: CIRCUIT_URLS.index,
             method: 'GET',
             handler: function (_, reply) {
                 reply.send({
@@ -45,7 +45,7 @@ test('should open circuit page after link click', async ({ page, server }) => {
             },
         },
         {
-            url: URLS.ref(circuit.ref),
+            url: CIRCUIT_URLS.ref(circuit.ref),
             method: 'GET',
             handler: function (_, reply) {
                 reply.send(circuit);
@@ -62,7 +62,7 @@ test('should open circuit page after link click', async ({ page, server }) => {
 
 test('should render breadcrumbs correctly', async ({ page, server }) => {
     await setupServer(server, {
-        url: URLS.index,
+        url: CIRCUIT_URLS.index,
         method: 'GET',
         handler: function (_, reply) {
             reply.send({
