+++ packages/client/__tests__/unit/app/circuits/page.test.tsx
@@ -4,7 +4,7 @@ import { axios } from '~shared/api/axios';
 import axiosMockAdapter from 'axios-mock-adapter';
 
 import CircuitsPage from '~app/circuits/page';
-import { URLS } from '~entities/circuit/api/urls';
+import { CIRCUIT_URLS } from '~entities/circuit/api';
 
 import { CircuitsMock } from '~mocks/entities/circuit/Circuit.mock';
 import { RouterMock } from '~tests-utils/router/Router.mock';
@@ -15,7 +15,7 @@ const MockAdapter = new axiosMockAdapter(axios);
 
 describe('Circuits page', () => {
     it('should render correctly', async () => {
-        MockAdapter.onGet(URLS.index).replyOnce(200, {
+        MockAdapter.onGet(CIRCUIT_URLS.index).replyOnce(200, {
             data: CircuitsMock,
             count: CircuitsMock.length,
         });
@@ -32,7 +32,7 @@ describe('Circuits page', () => {
     });
 
     it('should render correctly without searchParams', async () => {
-        MockAdapter.onGet(URLS.index).replyOnce(200, {
+        MockAdapter.onGet(CIRCUIT_URLS.index).replyOnce(200, {
             data: CircuitsMock,
             count: CircuitsMock.length,
         });
@@ -49,7 +49,7 @@ describe('Circuits page', () => {
     });
 
     it('should render breadcrumbs correctly', async () => {
-        MockAdapter.onGet(URLS.index).replyOnce(200, {
+        MockAdapter.onGet(CIRCUIT_URLS.index).replyOnce(200, {
             data: [],
             count: 0,
         });
