+++ packages/client/__tests__/unit/entities/circuit/api/useCircuitServer.test.ts
@@ -1,7 +1,7 @@
 import { axios } from '~shared/api/axios';
 import axiosMockAdapter from 'axios-mock-adapter';
 
-import { URLS } from '~entities/circuit/api/urls';
+import { CIRCUIT_URLS } from '~entities/circuit/api';
 import { useCircuitServer } from '~entities/circuit/api';
 
 import { CircuitsMock } from '~mocks/entities/circuit/Circuit.mock';
@@ -13,14 +13,14 @@ describe('useCircuitServer', () => {
     it('should fetch circuit correct', async () => {
         const circuitMock = CircuitsMock[0];
 
-        MockAdapter.onGet(URLS.ref(circuitMock.ref)).replyOnce(200, circuitMock);
+        MockAdapter.onGet(CIRCUIT_URLS.ref(circuitMock.ref)).replyOnce(200, circuitMock);
 
         const { circuit } = await useCircuitServer(circuitMock.ref);
 
         expect(circuit).toEqual(circuitMock);
     });
     it('should catch an error and throw next', async () => {
-        MockAdapter.onGet(URLS.ref('')).networkErrorOnce();
+        MockAdapter.onGet(CIRCUIT_URLS.ref('')).networkErrorOnce();
 
         try {
             await useCircuitServer('');
