+++ packages/client/__tests__/unit/entities/circuit/api/useCircuitsServer.test.ts
@@ -1,7 +1,7 @@
 import { axios } from '~shared/api/axios';
 import axiosMockAdapter from 'axios-mock-adapter';
 
-import { URLS } from '~entities/circuit/api/urls';
+import { CIRCUIT_URLS } from '~entities/circuit/api';
 import { useCircuitsServer } from '~entities/circuit/api';
 
 import { CircuitsMock } from '~mocks/entities/circuit/Circuit.mock';
@@ -15,7 +15,7 @@ describe('useCircuitsServer', () => {
         const page = 1;
         const perPage = 1;
 
-        MockAdapter.onGet(URLS.index).replyOnce(200, {
+        MockAdapter.onGet(CIRCUIT_URLS.index).replyOnce(200, {
             data: circuitsFetch,
             count: CircuitsMock.length,
         });
@@ -30,7 +30,7 @@ describe('useCircuitsServer', () => {
         expect(circuitsData).toEqual(circuitsDataExpected);
     });
     it('should throw an exception next correctly', async () => {
-        MockAdapter.onGet(URLS.index).networkErrorOnce();
+        MockAdapter.onGet(CIRCUIT_URLS.index).networkErrorOnce();
 
         try {
             await useCircuitsServer({ page: 1, perPage: 1 });
