+++ packages/client/src/entities/circuit/api/index.ts
@@ -1,3 +1,3 @@
-export { URLS as CIRCUITS_URLS } from './urls';
+export { URLS as CIRCUIT_URLS } from './urls';
 export { useCircuitServer } from './useCircuitServer';
 export { useCircuitsServer } from './useCircuitsServer';
