+++ packages/client/src/entities/circuit/ui/index.ts
@@ -1,0 +1,1 @@
+export { CircuitListCard } from './list-card';
