+++ packages/client/src/entities/circuit/ui/list-card/index.ts
@@ -1,0 +1,1 @@
+export { CircuitListCard } from './CircuitListCard';
